requirement.txt
fastapi
uvicorn
google-generativeai
langchain-tavily
markdown_pdf
python-dotenv
langchain-google-genai
python-multipart
langchain
langgraph
google-generativeai
markdown_pdf
markdown 
markdown2 
weasyprint
markdown_pdf
IPython
fpdf

